move admin home file
